+++ Models/ItemModel.cs
@@ -18,7 +18,7 @@ namespace CSharp_Inventory.Models
 
         public int Quantity { get; set; }
 
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
         public string? Description { get; set; }
     }
